Update API service with more endpoints
